add styleguidist and some examples
