+++ package.json
@@ -15,7 +15,9 @@
     "start": "nwb serve-react-demo --port=4000",
     "test": "nwb test-react",
     "test:coverage": "nwb test-react --coverage",
-    "test:watch": "nwb test-react --server"
+    "test:watch": "nwb test-react --server",
+    "styleguide": "styleguidist server",
+    "styleguide:build": "styleguidist build"
   },
   "dependencies": {
     "chroma-js": "^1.3.4",
@@ -27,6 +29,7 @@
     "react-feather": "^1.0.8",
     "react-onclickoutside": "^6.7.1",
     "react-show": "^2.0.4",
+    "react-styleguidist": "^6.2.4",
     "styled-components": "^3.1.3",
     "styled-system": "^1.1.1"
   },
@@ -34,6 +37,7 @@
     "react": ">=15"
   },
   "devDependencies": {
+    "create-nwb-webpack-config": "^1.0.0",
     "markdown-loader": "^2.0.1",
     "nwb": "^0.21.5",
     "raw-loader": "^0.5.1",
