+++ src/components/Checkbox/Readme.md
@@ -1,0 +1,73 @@
+Checkbox sizes:
+
+```jsx
+<div>
+  <Checkbox
+    size="sm"
+    mr={1}
+    value="sm"
+    checked={state.sm}
+    onChange={(e, value, checked) => setState({ [value]: checked })}
+  />
+  <Checkbox
+    size="md"
+    mr={1}
+    value="md"
+    checked={state.md}
+    onChange={(e, value, checked) => setState({ [value]: checked })}
+  />
+  <Checkbox
+    size="lg"
+    mr={1}
+    value="lg"
+    checked={state.lg}
+    onChange={(e, value, checked) => setState({ [value]: checked })}
+  />
+  <Checkbox
+    size="xl"
+    mr={1}
+    value="xl"
+    checked={state.xl}
+    onChange={(e, value, checked) => setState({ [value]: checked })}
+  />
+</div>
+```
+
+Checkbox colors:
+
+```jsx
+<div>
+  <Checkbox
+    palette="primary"
+    label="primary"
+    mr={1}
+    value="primary"
+    checked={state.primary}
+    onChange={(e, value, checked) => setState({ [value]: checked })}
+  />
+    <Checkbox
+    palette="warning"
+    label="warning"
+    mr={1}
+    value="warning"
+    checked={state.warning}
+    onChange={(e, value, checked) => setState({ [value]: checked })}
+  />
+  <Checkbox
+    palette="danger"
+    label="danger"
+    mr={1}
+    value="danger"
+    checked={state.danger}
+    onChange={(e, value, checked) => setState({ [value]: checked })}
+  />
+  <Checkbox
+    palette="success"
+    label="success"
+    mr={1}
+    value="success"
+    checked={state.success}
+    onChange={(e, value, checked) => setState({ [value]: checked })}
+  />
+</div>
+```
