+++ src/components/Icon/Readme.md
@@ -1,0 +1,23 @@
+`react-interface` imports and exposes a series of svg icons powered by feathers, view all options at
+[https://feathericons.com](feathericons.com).
+
+```jsx
+<div style={{ display: 'flex', justifyContent: 'space-around' }}>
+  <Icon type="camera" color='primary' />
+  <Icon type="activity" color='primary' />
+  <Icon type="bell" color='primary' />
+  <Icon type="cloud" color='primary' />
+  <Icon type="anchor" color='primary' />
+  <Icon type="battery" color='primary' />
+  <Icon type="box" color='primary' />
+  <Icon type="cast" color='primary' />
+  <Icon type="bar-chart" color='primary' />
+  <Icon type="activity" color='primary' />
+  <Icon type="alert-circle" color='primary' />
+  <Icon type="aperture" color='primary' />
+  <Icon type="align-left" color='primary' />
+  <Icon type="check" color='primary' />
+  <Icon type="chevron-down" color='primary' />
+  <Icon type="copy" color='primary'  />
+</div>
+```
