+++ src/components/Select/Readme.md
@@ -1,0 +1,12 @@
+```jsx
+<Select
+  placeholder="Select an option"
+  options={[
+    { label: 'First Option', value: 'first' },
+    { label: 'Second Option', value: 'second' },
+    { label: 'Third Option', value: 'third' }
+  ]}
+  onChange={value => setState({ value })}
+  value={state.value}
+/>
+```
