+++ styleguide.config.js
@@ -1,0 +1,12 @@
+var path = require('path');
+var createNwbWebpackConfig = require('create-nwb-webpack-config');
+
+module.exports = {
+  webpackConfig: createNwbWebpackConfig(),
+  components: 'src/**/[A-Z]*.js',
+  ignore: ['**/components/**/\*Wrapper\.js'],
+  skipComponentsWithoutExample: true,
+  styleguideComponents: {
+    Wrapper: path.join(__dirname, 'src/styleguide/ThemeWrapper')
+  }
+};
